Move fake-plastic channel into go/vumitools/channel and add a channel store for it.
