+++ go/vumitools/channel/__init__.py
@@ -1,0 +1,3 @@
+from go.vumitools.channel.models import ChannelStore
+
+__all__ = ['ChannelStore']
